refactor camera driver hw test
