+++ packages/camera_driver/include/hardware_test_camera/__init__.py
@@ -1,0 +1,1 @@
+from .hardware_test_camera import HWTestCamera
